pyproject.toml
[tool.poetry]
name = "hoechstgan"
version = "0.1.0"
description = ""
authors = ["Georg Wölflein <georgw7777@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.22.3"
torch = "^1.11.0"
torchvision = "^0.12.0"
matplotlib = "^3.5.1"
wandb = "^0.12.15"
scipy = "^1.8.0"
hydra-core = "^1.1.2"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
